+++ include/lib.cpp
@@ -235,19 +235,19 @@ map < pair < int, int >, pair < int, int > > lib::getCoordTranslationTable ( ) {
 
 }
 
-long double lib::phi ( long double v ) {
+double lib::phi ( double v ) {
 
 	return pow ( 1.0L + exp ( -v ), -1.0L );
 
 }
 
-vector < vector < long double > > lib::matrixMultiplication ( vector < vector < long double > > & v1, vector < vector < long double > > & v2 ) {
+vector < vector < double > > lib::matrixMultiplication ( vector < vector < double > > & v1, vector < vector < double > > & v2 ) {
 
-	vector < vector < long double > > result;
+	vector < vector < double > > result;
 
 	for ( size_t row = 0; row < v1.size ( ); ++ row ) {
 
-		result.push_back ( * new vector < long double > );
+		result.push_back ( * new vector < double > );
 
 		for ( size_t column = 0; column < v1.size ( ); ++ column ) {
 
@@ -261,9 +261,9 @@ vector < vector < long double > > lib::matrixMultiplication ( vector < vector < 
 
 }
 
-vector < long double > lib::matrixVectorMultiplication ( vector < vector < long double > > & v1, vector < long double > & v2 ) {
+vector < double > lib::matrixVectorMultiplication ( vector < vector < double > > & v1, vector < double > & v2 ) {
 
-	vector < long double > result ( v1.size ( ), 0 );
+	vector < double > result ( v1.size ( ), 0 );
 
 	for ( size_t i = 0; i < v1.size ( ); ++ i ) {
 
@@ -276,9 +276,9 @@ vector < long double > lib::matrixVectorMultiplication ( vector < vector < long 
 
 }
 
-vector < long double > lib::vectorPairMul ( vector < long double > & v1, vector < long double > & v2 ) {
+vector < double > lib::vectorPairMul ( vector < double > & v1, vector < double > & v2 ) {
 
-	vector < long double > result;
+	vector < double > result;
 
 	for ( size_t i = 0; i < v1.size ( ) && i < v2.size ( ); ++ i )
 		result.push_back ( v1 [ i ] * v2 [ i ] );
@@ -287,13 +287,13 @@ vector < long double > lib::vectorPairMul ( vector < long double > & v1, vector 
 
 }
 
-vector < vector < long double > > lib::matrixAdd ( vector < vector < long double > > & v1, vector < vector < long double > > & v2 ) {
+vector < vector < double > > lib::matrixAdd ( vector < vector < double > > & v1, vector < vector < double > > & v2 ) {
 
-	vector < vector < long double > > result;
+	vector < vector < double > > result;
 
 	for ( size_t i = 0; i < v1.size ( ) && i < v2.size ( ); ++ i ) {
 
-		result.push_back ( * new vector < long double > );
+		result.push_back ( * new vector < double > );
 
 		for ( size_t j = 0; j < v1 [ i ].size ( ) && j < v2 [ i ].size ( ); ++ j )
 			result [ i ].push_back ( v1 [ i ] [ j ] + v2 [ i ] [ j ] );
@@ -305,9 +305,9 @@ vector < vector < long double > > lib::matrixAdd ( vector < vector < long double
 }
 
 
-vector < vector < long double > > lib::matrixMulCoefficient ( long double k, vector < vector < long double > > & v ) {
+vector < vector < double > > lib::matrixMulCoefficient ( double k, vector < vector < double > > & v ) {
 
-	vector < vector < long double > > result ( v.size ( ), * new vector < long double > );
+	vector < vector < double > > result ( v.size ( ), * new vector < double > );
 
 	for ( size_t i = 0; i < v.size ( ); ++ i )
 
@@ -318,13 +318,13 @@ vector < vector < long double > > lib::matrixMulCoefficient ( long double k, vec
 
 }
 
-vector < vector < long double > > lib::transpose ( vector < vector < long double > > & v ) {
+vector < vector < double > > lib::transpose ( vector < vector < double > > & v ) {
 
-	vector < vector < long double > > result;
+	vector < vector < double > > result;
 
 	if ( v.size ( ) < 1 ) return result;
 
-	for ( size_t i = 0; i < v [ 0 ].size ( ); ++ i ) result.push_back ( * new vector < long double > );
+	for ( size_t i = 0; i < v [ 0 ].size ( ); ++ i ) result.push_back ( * new vector < double > );
 
 	for ( size_t i = 0; i < v.size ( ); ++ i ) {
 
@@ -337,9 +337,9 @@ vector < vector < long double > > lib::transpose ( vector < vector < long double
 
 }
 
-long double lib::scalarProduct ( vector < long double > & v1, vector < long double > & v2 ) {
+double lib::scalarProduct ( vector < double > & v1, vector < double > & v2 ) {
 
-	long double result = 0;
+	double result = 0;
 
 	for ( size_t i = 0; i < v1.size ( ) && i < v2.size ( ); ++ i )
 		result += v1 [ i ] * v2 [ i ];
@@ -348,13 +348,13 @@ long double lib::scalarProduct ( vector < long double > & v1, vector < long doub
 
 }
 
-vector < vector < long double > > lib::vectorsToMatrix ( vector < long double > & v1, vector < long double > & v2 ) {
+vector < vector < double > > lib::vectorsToMatrix ( vector < double > & v1, vector < double > & v2 ) {
 
-	 vector < vector < long double > > result;
+	 vector < vector < double > > result;
 
 	 for ( size_t i = 0; i < v1.size ( ); ++ i ) {
 
-		 result.push_back ( * new vector < long double > );
+		 result.push_back ( * new vector < double > );
 
 		 for ( size_t j = 0; j < v2.size ( ); ++ j )
 			result [ i ].push_back ( v1 [ i ] * v2 [ j ] );
@@ -365,17 +365,17 @@ vector < vector < long double > > lib::vectorsToMatrix ( vector < long double > 
 
 }
 
-vector < vector < vector < long double > > > lib::tensorAdd ( vector < vector < vector < long double > > > & v1, vector < vector < vector < long double > > > & v2 ) {
+vector < vector < vector < double > > > lib::tensorAdd ( vector < vector < vector < double > > > & v1, vector < vector < vector < double > > > & v2 ) {
 
-	vector < vector < vector < long double > > > result;
+	vector < vector < vector < double > > > result;
 
 	for ( size_t i = 0; i < v1.size ( ); ++ i ) {
 
-		result.push_back ( * new vector < vector < long double > > );
+		result.push_back ( * new vector < vector < double > > );
 
 		for ( size_t j = 0; j < v1 [ i ].size ( ); ++ j ) {
 
-			result [ i ].push_back ( * new vector < long double > );
+			result [ i ].push_back ( * new vector < double > );
 
 			for ( size_t k = 0; k < v1 [ i ] [ j ].size ( ); ++ k ) result [ i ] [ j ].push_back ( v1 [ i ] [ j ] [ k ] + v2 [ i ] [ j ] [ k ] );
 
@@ -387,9 +387,9 @@ vector < vector < vector < long double > > > lib::tensorAdd ( vector < vector < 
 
 }
 
-vector < pair < vector < long double >, vector < long double > > > lib::getDatasets ( istream & is ) {
+vector < pair < vector < double >, vector < double > > > lib::getDatasets ( istream & is ) {
 
-	vector < pair < vector < long double >, vector < long double > > > datasets;
+	vector < pair < vector < double >, vector < double > > > datasets;
 
 	int temp;
 	string in;
@@ -423,7 +423,7 @@ vector < pair < vector < long double >, vector < long double > > > lib::getDatas
 
 }
 
-void lib::printDatasets ( ostream & os, vector < pair < vector < long double >, vector < long double > > > datasets ) {
+void lib::printDatasets ( ostream & os, vector < pair < vector < double >, vector < double > > > datasets ) {
 
 	for ( size_t i = 0; i < datasets.size ( ); ++ i ) {
 
@@ -437,7 +437,7 @@ void lib::printDatasets ( ostream & os, vector < pair < vector < long double >, 
 
 }
 
-long double lib::intToIndata ( int i, int player ) {
+double lib::intToIndata ( int i, int player ) {
 
 	switch ( i ) {
